requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.8.2
South==1.0.2
argparse==1.2.1
django-imagefit==0.4
django-simple-captcha==0.4.5
psycopg2==2.5.4
six==1.9.0
wsgiref==0.1.2
